app/requirements.txt
python-dotenv
mercapi
discord-webhook
APScheduler
Flask